refactor(meter): ♻️  remove __examples__ to be consistent
