--- bundlesize.config.json
@@ -1,8 +1,0 @@
-{
-  "files": [
-    {
-      "path": "./dist/umd/renderless-components.min.js",
-      "maxSize": "50kb"
-    }
-  ]
-}
