+++ package.json
@@ -29,8 +29,7 @@
     "build-storybook": "build-storybook",
     "build:cjs": "cross-env BABEL_ENV=cjs babel src --extensions .ts,.tsx --config-file ./babel-config.js -d dist/cjs --source-maps",
     "build:esm": "cross-env BABEL_ENV=esm babel src --extensions .ts,.tsx --config-file ./babel-config.js -d dist/esm --source-maps",
-    "build:types": "tsc --emitDeclarationOnly --declaration --declarationDir dist/types",
-    "bundlesize": "bundlesize",
+    "build:types": "tsc --emitDeclarationOnly",
     "commit": "gacp",
     "format": "prettier --write \"./**/*.{js,ts,css,less,json,md,html,yml,yaml,pcss,jsx,tsx}\"",
     "keys": "node scripts/build/keys",
@@ -91,7 +90,6 @@
     "babel-loader": "^8.1.0",
     "babel-plugin-chakra-ui": "1.0.0-rc.5",
     "babel-plugin-date-fns": "^2.0.0",
-    "bundlesize": "0.18.0",
     "chalk": "4.1.0",
     "concurrently": "5.3.0",
     "conventional-github-releaser": "3.1.5",
