+++ tsconfig.json
@@ -3,30 +3,25 @@
     "target": "es5",
     "module": "esnext",
     "lib": ["dom", "esnext"],
-    "declaration": true,
-    "declarationDir": "dist/types",
-    "sourceMap": true,
-    "moduleResolution": "node",
-    "skipLibCheck": true,
-    "strict": false,
-    "noFallthroughCasesInSwitch": true,
-    "suppressImplicitAnyIndexErrors": true,
-    "noImplicitAny": true,
-    "strictFunctionTypes": true,
-    "strictNullChecks": true,
-    "strictPropertyInitialization": true,
     "jsx": "react",
     "esModuleInterop": true,
     "allowSyntheticDefaultImports": true,
+    "moduleResolution": "node",
+    "skipLibCheck": true,
     "downlevelIteration": true,
     "resolveJsonModule": true,
-    "types": ["node", "jest", "@testing-library/jest-dom"]
+    "strict": false,
+    "strictFunctionTypes": true,
+    "strictNullChecks": true,
+    "strictPropertyInitialization": true,
+    "suppressImplicitAnyIndexErrors": true,
+    "noImplicitAny": true,
+    "noFallthroughCasesInSwitch": true,
+    "types": ["node", "jest", "@testing-library/jest-dom"],
+    "sourceMap": true,
+    "declaration": true,
+    "declarationDir": "dist/types"
   },
-  "exclude": [
-    "node_modules",
-    "dist",
-    "**/*/stories",
-    "**/*/__tests__",
-    "**/*/__examples__"
-  ]
+  "include": ["src"],
+  "exclude": ["**/*/stories", "**/*/__tests__"]
 }
